rename config to project, flesh out some features
