--- cmd/config/doc.go
@@ -1,2 +1,0 @@
-// Package configcmd implements the 'inertia config' subcommands
-package configcmd
