+++ cmd/project/doc.go
@@ -1,0 +1,2 @@
+// Package projectcmd implements the 'inertia project' subcommands
+package projectcmd
