+++ local/init.go
@@ -2,12 +2,10 @@ package local
 
 import (
 	"errors"
-	"fmt"
 	"os"
 	"path/filepath"
 
 	"github.com/ubclaunchpad/inertia/cfg"
-	"github.com/ubclaunchpad/inertia/local/git"
 )
 
 // Init sets up Inertia configuration
@@ -21,21 +19,14 @@ func Init() (*cfg.Inertia, error) {
 	return inertia, Write(configPath, inertia)
 }
 
-// InitProject creates the inertia config folder and returns an error if we're
-// not in a git project.
-func InitProject(path, name string, defaultProfile cfg.Profile) error {
-	cwd, err := os.Getwd()
-	if err != nil {
-		return err
-	}
-
-	if err = git.IsRepo(cwd); err != nil {
-		return fmt.Errorf("could not find git repository: %s", err.Error())
-	}
+// InitProject creates the inertia config file and returns an error if Inertia
+// is already configured
+func InitProject(path, name, host string, defaultProfile cfg.Profile) error {
 	if s, _ := os.Stat(path); s != nil {
 		return errors.New("inertia is already properly configured in this directory")
 	}
-	var project = cfg.NewProject(name)
+
+	var project = cfg.NewProject(name, host)
 	project.SetProfile("default", defaultProfile)
 
 	return Write(path, project)
